requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
transformers==4.34.2
torch==2.0.1
accelerate==0.24.0
bitsandbytes==0.41.1
langchain==0.0.289
scipy==1.11.2
celery==5.3.4
redis==5.0.0
